Refactor callback registrar to avoid needing the `__init__` method to create the callback list
Previously, the CallbackRegistrarModuleMixin's constructor
directly initialized the _nshtrainer_callbacks list. This could
lead to potential conflicts if a subclass also defined an __init__
method. By converting _nshtrainer_callbacks to a property that
lazily initializes the underlying list on first access, we avoid
any unintentional overwriting of the callback list.

This change improves the robustness and reusability of the mixin.